requirements.txt
spoon-ai-sdk
flask
twilio
requests
python-dotenv
web3